random dice creation end of branch
